chapter-5-multimodal-ai/llm_integration/gemini_integration.py
"""
Google Gemini multimodal integration.
"""

from typing import List, Optional
import google.generativeai as genai
from PIL import Image


class GeminiMultimodal:
    """Gemini multimodal capabilities"""
    
    def __init__(self, api_key: str):
        genai.configure(api_key=api_key)
        self.model = genai.GenerativeModel('gemini-pro-vision')
    
    def analyze_image(self, 
                     image_path: str,
                     prompt: str) -> str:
        """
        Analyze image with Gemini
        
        Args:
            image_path: Path to image
            prompt: Analysis prompt
            
        Returns:
            Analysis result
        """
        image = Image.open(image_path)
        response = self.model.generate_content([prompt, image])
        return response.text
    
    def analyze_images_batch(self,
                            image_paths: List[str],
                            prompts: List[str]) -> List[str]:
        """
        Batch analyze multiple images
        
        Args:
            image_paths: List of image paths
            prompts: List of prompts (one per image)
            
        Returns:
            List of analysis results
        """
        results = []
        
        for image_path, prompt in zip(image_paths, prompts):
            result = self.analyze_image(image_path, prompt)
            results.append(result)
        
        return results
    
    def video_analysis(self,
                      video_path: str,
                      prompt: str) -> str:
        """
        Analyze video (experimental)
        
        Args:
            video_path: Path to video
            prompt: Analysis prompt
            
        Returns:
            Analysis result
        """
        # Note: Video analysis may require different model
        # This is a placeholder for future implementation
        raise NotImplementedError("Video analysis not yet implemented")