chapter-5-multimodal-ai/requirements.txt
# Core Dependencies
anthropic>=0.18.0
pillow>=10.0.0
numpy>=1.24.0

# Vision
torch>=2.0.0
torchvision>=0.15.0
clip @ git+https://github.com/openai/CLIP.git
transformers>=4.30.0
groundingdino @ git+https://github.com/IDEA-Research/GroundingDINO.git

# Audio
openai-whisper>=20230314
librosa>=0.10.0
soundfile>=0.12.0
gtts>=2.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10

# LLM Integrations
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
ipython>=8.14.0

# Optional: For specific features
# Uncomment if needed
# redis>=4.5.0  # For Redis message queue
# pika>=1.3.0  # For RabbitMQ
# scipy>=1.11.0  # For advanced audio processing
# scikit-learn>=1.3.0  # For ML utilities